[Pipelines] Fix MPNet bugs when running on GPU with bfloat16.
The pipelines CLI now runs on GPU by default if a GPU is present, and
was raising an error when running MPNet.

MAX_INTEGRATION_TESTS_REV_ID: 20760f8b129cf5cc9cb9239043882aa990fc28f8

+++ max/tests/integration/pipelines/python/test_common/torch_utils.py
@@ -110,7 +110,9 @@ def run_embeddings_generation(
             [prompt], padding=True, truncation=True, return_tensors="pt"
         ).to(device)
         output = model(**encoded_input)
-        embeddings = output.last_hidden_state.cpu().detach().numpy()
+        embeddings = (
+            output.last_hidden_state.cpu().detach().to(torch.float32).numpy()
+        )
         embeddings = embeddings[0]
         results.append({"prompt": prompt, "embeddings": embeddings})
     return results
